+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### BC breaks
 
   * raised minimum required PHP version to 7.1.0
+  * all methods of `org\bovigo\vfs\visitor\vfsStreamVisitor` are now declared with `self` as return type
 
 
 1.6.5 (2017-08-01)